nixos/top-level: fail more helpfully on forbidden dependencies
We now use nix why-depends in a terribly hacky way to show the paths
from the non-overridden toplevel to each forbidden dependency.

Note that this does not take into account dependencies from the build
closure, even if these are brought in using
system.includeBuildDependencies. I'm not sure if that's desirable or
not.