junicode: 2.100 -> 2.200
Release announcement:
https://github.com/psb1558/Junicode-font/releases/tag/v2.200

From it:

Junicode 2.200 features miscellaneous fixes and additions, and it also
inaugurates a long-term program of adding glyphs for the benefit of
the Ansund HTR (Handwritten Text Recognition) project, which is
developing tools for the automated reading of medieval manuscripts. Of
the additions made so far, the ones likely to be of greatest interest
to users are two series of medieval capitals, available via features
ss11[1] and ss11[2]. The first series consists of rustic capitals,
often used for text in late ancient and early medieval times and for
rubrics (headings) in the central Middle Ages. The second consists of
Lombardic capitals, used in the central and later Middle Ages for what
are now called drop caps. These capitals are designed to harmonize
with Junicode to the greatest degree possible while remaining faithful
to the medieval sources.