nixos-containers: allow customizing `--link-journal` option
`systemd-nspawn`'s `--link-journal` option allows storing journal on
the host or on the guest. Previously it was hardcoded to store the
container journal on the guest and symlink it to the host's
`/var/log`. However, for ephemeral containers it becomes impossible to
read the logs, and for systems using some sort of impersistence
mechanism it is impossible to store logs without persisting them using
`bindMounts`, which is essentially what `--link-journal={try-,}host`
does.

Allowing to customize this option brings the ability to persist logs
for ephemeral containers.