nixos/mosquitto: remove the bind_interface test
bind_interface is the mosquitto way of trying to bind to all addresses
on an interface, but it is unreliable (trying to bind to link-local v6
addresses *sometimes* but not always) and just prone to failure in
general for reasons we have yet to discover.

since this kind of automatic behavior isn't particularly necessary in a
declarative system we may as well skip it.