nixos-rebuild: use `--pipe` instead of `--pty`
Accordingly to `systemd-run` manpage, `--pty` is to be used by running
programs that expect interactive user input/output usage. This is not
the case of `switch-to-configuration` script.

One issue of using `--pty` is that it will fail in case the user is not
running the command inside a terminal, see issue #262686.

This commit changes `system-run` call to use `--pipe` flag instead.
`--pipe` will inherit the standard input/output/error from `systemd-run`
itself, instead of terminal like `--pty`. This should fix the issue of
running `nixos-rebuild switch` in non-terminal use cases.

We also add `--scope` flag (and remove `--service-type` since it would
be ignored), that will create a scoped service that will be inherited by
`systemd-run` itself instead of service manager.

Fix #262686.