python311Packages.keyutils: fix build
This fixes the build for keyutils on Python 3.11 and higher.

It appears this package is affected by changes that were made to CPython
that were causing code generated by certain earlier versions of Cython
to fail.

This update simply re-generates the relevant code with the up-to-date
Cython 0.29.x series during the pre-build phase, allowing the build to
succeed.

Fixes #263482.
Fixes #263506.

Co-authored-by: Robert Schütz <github@dotlambda.de>
Co-authored-by: Martin Weinelt <mweinelt@users.noreply.github.com>