nsncd: unstable-2022-11-14 -> unstable-2023-10-26
Note: we decided to rewrite the history of the fork who somehow got
out of hand. Feature-wise, this version bump fixes the various host
faulty behaviour. See the
https://github.com/nix-community/nsncd/pull/9 and
https://github.com/nix-community/nsncd/pull/10 PRs for more details.

We're in the process of upstreaming this change to twosigma/nsncd,
however, upstream has been pretty slow to review our PRs so far. Since
the hostname bug surfaces quite regularly in the Nixpkgs issue
tracker, we decided to use the nix-community fork as canon for Nixpkgs
for now.

Fixes: https://github.com/NixOS/nixpkgs/issues/132646
Fixes: https://github.com/NixOS/nixpkgs/issues/261269