substitue-all-files: Change the separator from \0 to \n.
Using \0 separator with a list of files trigger a bug when file names start with '0'. \0 is interpreted like \00 and instead of passing a list of files ot xargs, a string with a concanetation of all file names concatenated with \00 is piped to xargs.

Using \n as a separator solves this issue

Changes to be committed:
  modified:   pkgs/build-support/substitute-files/substitute-all-files.nix