buildRustCrate: fix cross
Previously, `cargo` / `rustc` were run while letting `cc-rs` guess what should be the linker, the cc, etc.

In general, this works almost well, except when `cc-rs` is confused and does not.

As we already do in many places in the Rust cross compilation ecosystem, we set certain environment variables
which wraps any rustc or cargo invocation to help `cc-rs` or `cargo` understand what is the linker, etc.

We proceed to do the same here.