llvmPackages_7.clang: use the stdenv libc++ on Darwin
This change treats the top-level libc++ as the “system” libc++ on
Darwin, matching the way libstdc++ is used on Linux. The approach it
takes is to use the requested version’s libc++ headers while linking
against the top-level library. This is safe to do because libc++ is
built with the stable ABI on Darwin by default.

This change is necessary to support building a package with one version
of clang while linking against libraries that are built with different
versions. Otherwise, if they both use C++, they will link against
different versions of libc++ and (probably) crash at runtime. Examples of
where this is an issue include older versions of nodejs building with
clang 15 and darktable (which is currently built with clang 13).