freetube: use electron_27 instead of electron_22
Upstream freetube-19.1 uses electron-27, so match that.

This fixes the issue that electron_22 is marked as insecure and results
in eval error by default. And with e.g. the default electron in the
stable branch (electron_26) we hit
https://github.com/FreeTubeApp/FreeTube/issues/3953 ("App turns entirely
white a few moments after startup on wayland-WM-only-setups").

This patch is only for the stable branch, as unstable has electron_27 as
default and unpinned electron in freetube.