glom: provide bin dir of postgresql
By default `pg_config --bindir` is executed during configure phase to
get the bin directory containing all the postgres executables used by
glom.

This seems wrong given that the same postgresql_15 package will be
executed on build and on the final host platform which is not
necessarily the same. To avoid that, specify the bin dir manually via
the corresponding configure option.