lib.systems.inspect: deprecate isEfi
This predicate never did what it claimed.

(U)EFI support is a property of a specific motherboard, not of a
platform or an entire architecture.  This predicate was used only by
two packages.  This commit relocates the data previously held by
`inspect.isEfi` to the `meta.platforms` of `gnu-efi` and the
`withEfi` default value of `systemd`.

We need to make room for `isEfi` (or perhaps `isUefi`) in
preparation for https://github.com/NixOS/nixpkgs/pull/231951 so we
should start the deprecation process sooner rather than later.