llvmPackages_12.openmp,llvmPackages_13.openmp: fix cross
CMake commit faa950a155 ("try_compile: Run native build tool with verbose output")
did not play well with these versions of openmp, which have a faulty
failure regex that considers "unknown" in the build output to indicate
a failure.  When cross compiling, the string "unknown" is very likely
to occur as part of triples.  Fix by backporting a patch that improves
the failure regex check to not be tripped up by triples.

The same problem affects LLVM 11 and possibly earlier, but the patch
doesn't apply that far back, so I didn't bother for now.

Fixes: b8ac5b6cfbc7 ("cmake: 3.25.3 -> 3.26.4")