abseil-cpp: don’t propagate CoreFoundation
abseil-cpp builds with the default SDK, which can break dependent
packages that require a different SDK version to build. Don’t propagate
the CoreFoundation framework path and rely on nixpkgs to provide it
instead. The line can’t be dropped because otherwise abseil-cpp will
fail to build (due to missing symbols).