nixos/unpoller: allow host paths for password files
In the current module, the type checker restricts password files to paths;
those are then coerced to strings (the resulting paths in the Nix store) and
plugged into the unpoller config file as `file://${s}`. This isn't
compatible with secret provisioning schemes (e.g. sops) that avoid having
secrets in the Nix store.

Just switching the type to str avoids the problem. Configs that use paths
will continue to work transparently via the coercion from path to str, and
users that prefer provisioning secrets to paths on the host can do so.