qt5.qtbase: ignore hook propagated from buildPlatform by comparing stdenvs
The main headache with cross-compiling QT is that our QT derivations
go bananas with propagatedInputs.  Unfortunately this means that
both buildPackages.qtbase and hostPackages.qtbase get propagated
into the dependencies of every cross-compiled builds, which triggers
the hook check.

Previously in https://github.com/NixOS/nixpkgs/pull/227900 I simply
disabled the check outright for cross-compiled builds.  This is a
more refined solution, where we skip the hook if it propagates
across a platform change.