nixos/tests/systemd-networkd-dhcpserver-static-leases: fix build
Before this patch, `192.168.1.1` was assigned to `eth1` of `client` and
DHCP was turned off (`null` is the default for
`networking.interfaces.<name>.useDHCP` which means that `DHCP=no` is set
if an IP is configured statically). Because of that, the static
10.0.0.10 was never assigned.

The VLAN 1 (with `eth1` in it) never gets another IP assigned statically
now to make sure `systemd-networkd-wait-online` doesn't succeed before
10.0.0.10 was configured via DHCP.

DHCP is turned off globally (as it was before) because this means that
`--any` is not set for `systemd-networkd-wait-online` and as a result
it's explicitly waited for `eth1` to be ready before checking its IPs
(otherwise that would happen too early because `eth0` reaches
`routable`/`configured` quicker).