geany: 1.38 -> 2.0
https://www.geany.org/documentation/releasenotes/2.0/

Upstream now supports building via Meson, which is declared experimental
at the time of writing; Autotools is still the default. [1]

The correct license for Geany is GPL-2.0-or-later, as indicated by [2].

Add a patch that disables `test_sidebar` because it runs into undefined
behavior in headless environments, and crashes at least on headless
Darwin. The patch can be removed if [3] is merged (or the issue fixed
otherwise).

[1] https://github.com/geany/geany/pull/2761/commits
[2] https://github.com/geany/geany/tree/2.0.0#license
[3] https://github.com/geany/geany/pull/3676