Revert "spirv-llvm-translator: move spirv-tools to nativeBuildInputs"
This reverts commit 69a71a0ec2da81a4213e1f0ceedb95545d9e2414.

Previously, the only use for spirv-tools was its commands being used
to run tests, but now, spirv-llvm-translator also links against its
libraries.  Tests won't be run when cross compiling to a
non-compatible architecture anyway, so to keep spirv-llvm-translator
cross compiling, we should move spirv-tools back to being a build
input so that the libraries can be found.  We could try to convince
CMake to use SPIRV-Tools.cmake from the host spirv-tools and
SPIRV-Tools-tools.cmake from the build spirv-tools, but since we never
actually need both when cross compiling it's not worth it.

This fixes the cross build LLVM/SPIRV-LLVM-Translator 16.