deviceTree: ensure file symlinks are included in applyOverlays output
When using a custom `hardware.deviceTree.dtbSource`, we cannot assume
that all DTBs in this directory are regular files. This change allows
for file symlinks to be present as well, which fixes the issue where
only file (a symlink) is present in `dtbSource` and the copy fails with
`cp: missing file operand`.