python3Packages.editableInstallHook: init
Serves as a replacement for the old and busted `setuptoolsShellHook`. It
needs to be added manually to reduce the magic factor.

Supports `pyproject.toml`-based projects in addition to `setup.py`.
Always runs (vs. detecting the existence of these files) to reduce the
magic factor. If no suitable project is found, then a big red error
message is displayed by `pip` but shell initialization nonetheless
continues as normal.

Documentation is updated to describe existence and usage.