ubootTools: fix cross properly this time
When we're building a ubootTools, we want to produce a set of tools that runs
on the host, so we want their dependencies to be built for hostPlatform.

When we're building an uboot, the tools will run on the builder as part
of the build, so we want their dependencies to be built for buildPlatform.