poppler: remove unused pcre
Poppler does not directly depend on PCRE.

It was added in <https://github.com/NixOS/nixpkgs/pull/127832>,
presumably to suppress CMake complaining about `glib-2.0` needing `pcre`
because CMake is not aware that we interpret `Requires.private`
field in pkg-config files differently.
And GLib since switched to pcre2 anyway so the warning is back.