vesktop: use a single store for all supported platforms
pnpm 8.10.0 introduced a feature that allows downloading npm deps for
other platforms. Sadly, this is a package.json option currently, but
using a simple patch with jq we can leverage this feature, to have a
single output hash for all supported platforms for our pnpm
dependencies. This might be one step forward towards an eventual
fetchPnpmDeps build-support function.

Signed-off-by: Sefa Eyeoglu <contact@scrumplex.net>