lib.licenses.parity70: mark non-free
This licenses forces licensees to share changes made for personal use.
I think this makes this a non-free license.

As part of its license review process, Debian applies the Desert
Island and Dissident tests[1], both of which this license fails as far as
I can tell.

AFAICT, It hasn't been reviewed by any of the usual suspects (Debian,
Fedora, FSF), but if any of them do make a determination on it, we can
take that into account and update our own view.

[1]: https://people.debian.org/~bap/dfsg-faq.html#testing